requirements.txt
azureml-sdk
pyyaml
pandas
sklearn